Modified files little bit more as per class lectures
