+++ ex.py
@@ -12,23 +12,12 @@ hand-written digits, from 0-9.
 # License: BSD 3 clause
 
 # Standard scientific Python imports
-import matplotlib.pyplot as plt
+#import matplotlib.pyplot as plt
 
 # Import datasets, classifiers and performance metrics
 from sklearn import datasets, metrics, svm
 from sklearn.model_selection import train_test_split
-
-
-def split_train_dev_test(X, y, test_size, dev_size):
-    train_size = 1 - dev_size - test_size
-    X_train, X_rem, y_train, y_rem = train_test_split(X, y, train_size=train_size, shuffle=False)
-    new_test_size = test_size / (1 - train_size)    
-    X_dev, X_test, y_dev, y_test = train_test_split(X_rem, y_rem, test_size=new_test_size, shuffle=False)  
-    return X_train, X_dev, X_test, y_train, y_dev, y_test
-
-def predict_and_eval(model, X_test, y_test):
-    predicted = model.predict(X_test)
-    return predicted
+from utils import split_train_dev_test, predict_and_eval, train_model
 
 
 ###############################################################################
@@ -47,11 +36,11 @@ def predict_and_eval(model, X_test, y_test):
 
 digits = datasets.load_digits()
 
-_, axes = plt.subplots(nrows=1, ncols=4, figsize=(10, 3))
-for ax, image, label in zip(axes, digits.images, digits.target):
-    ax.set_axis_off()
-    ax.imshow(image, cmap=plt.cm.gray_r, interpolation="nearest")
-    ax.set_title("Training: %i" % label)
+# _, axes = plt.subplots(nrows=1, ncols=4, figsize=(10, 3))
+# for ax, image, label in zip(axes, digits.images, digits.target):
+#     ax.set_axis_off()
+#     ax.imshow(image, cmap=plt.cm.gray_r, interpolation="nearest")
+#     ax.set_title("Training: %i" % label)
 
 ###############################################################################
 # Classification
@@ -72,71 +61,26 @@ for ax, image, label in zip(axes, digits.images, digits.target):
 n_samples = len(digits.images)
 data = digits.images.reshape((n_samples, -1))
 
-# Create a classifier: a support vector classifier
-clf = svm.SVC(gamma=0.001)
 
-# Split data into 50% train and 50% test subsets
+# Split data into 50% train and 30% test and 20% Validate subsets
 X_train, X_Dev, X_test, y_train, y_dev, y_test  = split_train_dev_test(
-    data, digits.target, test_size=0.1, dev_size=0.1
+    data, digits.target, test_size=0.3, dev_size=0.2
 )
 
-# Learn the digits on the train subset
-clf.fit(X_train, y_train)
+clf = train_model(X_train, y_train, {'gamma':0.001})
 
 # Predict the value of the digit on the test subset
-#predicted = clf.predict(X_test)
-predicted = predict_and_eval(clf, X_test, y_test)
+predict_and_eval(clf, X_test, y_test)
 
 ###############################################################################
 # Below we visualize the first 4 test samples and show their predicted
 # digit value in the title.
 
-_, axes = plt.subplots(nrows=1, ncols=4, figsize=(10, 3))
-for ax, image, prediction in zip(axes, X_test, predicted):
-    ax.set_axis_off()
-    image = image.reshape(8, 8)
-    ax.imshow(image, cmap=plt.cm.gray_r, interpolation="nearest")
-    ax.set_title(f"Prediction: {prediction}")
+# _, axes = plt.subplots(nrows=1, ncols=4, figsize=(10, 3))
+# for ax, image, prediction in zip(axes, X_test, predicted):
+#     ax.set_axis_off()
+#     image = image.reshape(8, 8)
+#     ax.imshow(image, cmap=plt.cm.gray_r, interpolation="nearest")
+#     ax.set_title(f"Prediction: {prediction}")
 
 ###############################################################################
-# :func:`~sklearn.metrics.classification_report` builds a text report showing
-# the main classification metrics.
-
-print(
-    f"Classification report for classifier {clf}:\n"
-    f"{metrics.classification_report(y_test, predicted)}\n"
-)
-
-###############################################################################
-# We can also plot a :ref:`confusion matrix <confusion_matrix>` of the
-# true digit values and the predicted digit values.
-
-disp = metrics.ConfusionMatrixDisplay.from_predictions(y_test, predicted)
-disp.figure_.suptitle("Confusion Matrix")
-print(f"Confusion matrix:\n{disp.confusion_matrix}")
-
-plt.show()
-
-###############################################################################
-# If the results from evaluating a classifier are stored in the form of a
-# :ref:`confusion matrix <confusion_matrix>` and not in terms of `y_true` and
-# `y_pred`, one can still build a :func:`~sklearn.metrics.classification_report`
-# as follows:
-
-
-# The ground truth and predicted lists
-y_true = []
-y_pred = []
-cm = disp.confusion_matrix
-
-# For each cell in the confusion matrix, add the corresponding ground truths
-# and predictions to the lists
-for gt in range(len(cm)):
-    for pred in range(len(cm)):
-        y_true += [gt] * cm[gt][pred]
-        y_pred += [pred] * cm[gt][pred]
-
-print(
-    "Classification report rebuilt from confusion matrix:\n"
-    f"{metrics.classification_report(y_true, y_pred)}\n"
-)
